move ClientList to Wire.API.User.Client
