+++ libs/wire-api/src/Wire/API/Conversation.hs
@@ -47,9 +47,6 @@ module Wire.API.Conversation
     GroupId (..),
     mlsSelfConvId,
 
-    -- * List of MLS client ids
-    ClientList (..),
-
     -- * Conversation properties
     Access (..),
     AccessRole (..),
@@ -234,19 +231,6 @@ instance ToSchema (Versioned 'V2 ConversationMetadata) where
         .= object
           "ConversationMetadata"
           (conversationMetadataObjectSchema accessRolesSchemaV2)
-
--- | Client list for internal API.
-data ClientList = ClientList {clClients :: [ClientId]}
-  deriving stock (Eq, Show, Generic)
-  deriving (Arbitrary) via (GenericUniform ClientList)
-  deriving (FromJSON, ToJSON, S.ToSchema) via Schema ClientList
-
-instance ToSchema ClientList where
-  schema =
-    object "ClientList" $
-      ClientList
-        <$> clClients
-          .= field "client_ids" (array schema)
 
 -- | Public-facing conversation type. Represents information that a
 -- particular user is allowed to see.
